requirements.txt
asgiref==3.2.7
astroid==2.4.2
azure-common==1.1.25
azure-core==1.6.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==3.0.7
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
idna==2.9
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
msrest==0.6.16
oauthlib==3.1.0
pycparser==2.20
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
